// src/shared/api/generated/models/eventCatalogRevisionResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { EventCatalogRevisionResponseDtoCompatibility } from "./eventCatalogRevisionResponseDtoCompatibility";
import type { EventCatalogRevisionResponseDtoLifecycle } from "./eventCatalogRevisionResponseDtoLifecycle";
import type { EventCatalogRevisionResponseDtoOrigin } from "./eventCatalogRevisionResponseDtoOrigin";
import type { EventCatalogRevisionResponseDtoPayloadSchema } from "./eventCatalogRevisionResponseDtoPayloadSchema";

export interface EventCatalogRevisionResponseDto {
  /** @nullable */
  archivedAt?: string | null;
  clientIngestible: boolean;
  code: string;
  compatibility: EventCatalogRevisionResponseDtoCompatibility;
  countsAsActivity: boolean;
  createdAt: string;
  /** @nullable */
  currentRevisionId?: string | null;
  definitionKeyId: string;
  /** @nullable */
  description?: string | null;
  enabled: boolean;
  id: string;
  isCurrent: boolean;
  lifecycle: EventCatalogRevisionResponseDtoLifecycle;
  metadataUpdatedAt: string;
  name: string;
  origin: EventCatalogRevisionResponseDtoOrigin;
  payloadSchema: EventCatalogRevisionResponseDtoPayloadSchema;
  /** @minimum 0 */
  pinnedScenarioRevisionCount: number;
  policyUpdatedAt: string;
  /** @minimum 1 */
  policyVersion: number;
  projectId: string;
  readOnly: boolean;
  updatedAt: string;
  /** @minimum 1 */
  version: number;
}